requirements.txt
dspy-ai
psutil
torch
google-generativeai